+++ homematicip/aio/class_maps.py
@@ -9,6 +9,7 @@ TYPE_CLASS_MAP = {
     DeviceType.ALARM_SIREN_INDOOR: AsyncAlarmSirenIndoor,
     DeviceType.BRAND_DIMMER: AsyncBrandDimmer,
     DeviceType.BRAND_SHUTTER: AsyncFullFlushShutter,
+    DeviceType.BRAND_PUSH_BUTTON: AsyncBrandPushButton,
     DeviceType.BRAND_SWITCH_MEASURING: AsyncBrandSwitchMeasuring,
     DeviceType.BRAND_SWITCH_NOTIFICATION_LIGHT: AsyncBrandSwitchNotificationLight,
     DeviceType.BRAND_WALL_MOUNTED_THERMOSTAT: AsyncWallMountedThermostatPro,