+++ homematicip/aio/device.py
@@ -182,6 +182,8 @@ class AsyncPushButton(PushButton, AsyncDevice):
 class AsyncPushButton6(PushButton6, AsyncPushButton):
     """ HMIP-WRC6 (Wall-mount Remote Control - 6-button)  """
 
+class AsyncBrandPushButton(BrandPushButton, AsyncPushButton):
+    """ HMIP-BRC2 (Remote Control for brand switches – 2x channels) """
 
 class AsyncKeyRemoteControl4(KeyRemoteControl4, AsyncPushButton):
     """ HMIP-KRC4 (Key Ring Remote Control - 4 buttons) """